+++ platform/fpga/openpiton/platform.c
@@ -143,7 +143,7 @@ static int openpiton_irqchip_init(bool cold_boot)
 			return ret;
 	}
 
-	return plic_warm_irqchip_init(&plic);
+	return plic_warm_irqchip_init();
 }
 
 /*
